initial implementation for parameters abstraction

+++ Include/IStrategizer/Vector2.h
@@ -24,15 +24,15 @@ namespace IStrategizer
             return this->X != p_right.X || this->Y == p_right.Y;
         }
 
-        double Distance (const Vector2& p_other)
+        int Distance (const Vector2& p_other)
         {
-            return sqrt((pow((double)(p_other.X - this->X), 2) + pow((double)(p_other.Y - this->Y), 2)));
+            return (int)sqrt((pow((double)(p_other.X - this->X), 2) + pow((double)(p_other.Y - this->Y), 2)));
         }
 
-        bool IsNull() { return X == -1 && Y == -1; }
-        static const Vector2& Zero() { static Vector2 zero; return zero; }
-        static const Vector2& One() { static Vector2 one(1, 1); return one; }
-        static const Vector2& Null() { static Vector2 one(-1, -1); return one; }
+        bool IsNull()                    { return X == -1 && Y == -1; }
+        static const Vector2& Zero()    { static Vector2 zero; return zero; }
+        static const Vector2& One()        { static Vector2 one(1, 1); return one; }
+        static const Vector2& Null()    { static Vector2 one(-1, -1); return one; }
 
     };
 }
